Componentize the ChatSystem
to make it modular, and to make the chat page more customizable.
