+++ client/src/containers/DirectMessages.js
@@ -1,28 +1,7 @@
 import Gun from 'gun';
 import React, { useReducer, useState } from 'react';
-import ChatInput from '../components/ChatInput';
-import ChatLog from '../components/ChatLog';
-import ChatSelector from '../components/ChatSelector';
+import { ChatSystem } from '../components/ChatSystem';
 import { MessageSender } from '../components/MessageSender';
-
-// MOCK DATA
-var conversations = [
-  {
-    id: "a",
-    messages: [
-      "Hello!",
-      "Goodbye!"
-    ]
-  },
-  {
-    id: "b",
-    messages: [
-      "Hola!",
-      "Adios!"
-    ]
-  }
-]
-// END MOCK DATA
 
 // initialize gun locally
 const gun = Gun();
@@ -48,8 +27,6 @@ function DirectMessages() {
     time: Date.now()
   })
 
-  const [currentConversation, setConversation] = useState(conversations[0]);
-
   // initialize the reducer & state for holding the messages array
   const [state/*, dispatch*/] = useReducer(reducer, initialState)
 
@@ -70,10 +47,6 @@ function DirectMessages() {
   //     })
   //   })
   // }, [])
-
-  function conversationChangeHandler(newConversationId) {
-    setConversation(conversations.find(e => e.id == newConversationId))
-  }
 
   // set a new message in gun, update the local state to reset the form field
   function sendCredits() {
@@ -101,11 +74,7 @@ function DirectMessages() {
   }
 
   return (
-    <div>
-      <ChatSelector onSelect={conversationChangeHandler} conversation={currentConversation}/>
-      <ChatLog conversation={currentConversation}/>
-      <ChatInput conversation={currentConversation}/>
-    </div>
+    <ChatSystem/>
   )
   
   // return (
